Started handling login verification, changed page destination in createAccount
There is now php code in checkLogin that will verify a user's information that they entered. If the information is corrent and a user exists with the matching credentials, the profile page is loaded. If there is no matching users, the page directs the user to create an arrount (does not direct them to the createAccount page). I also changed the destination of the createAccount page to be the profile page, instead of the newAccount page. After a user creates their account, we want to take them directly to their new profile page.